III-4475 - Move priceinfo mutations to actual component
